fixed models

+++ Accounting/Models/Depreciation.cs
@@ -5,6 +5,6 @@ namespace AccountingAPI.Models
     {
         public Guid FixedAssetId { get; set; }
         public Guid PeriodId { get; set; }
-        public double DepreciationAmount { get; set; }
+        public decimal DepreciationAmount { get; set; }
     }
 }
